feat: add linter configuration and refactor commands
- Delete obsolete `.DS_Store` file and add it to `.gitignore` along with `.idea/` and `bin/`
- Add `.golangci.yaml` for enabling and configuring several linters
- Remove `app` related cobra command files as part of a commands refactor
- Update the `cluster` cobra command handling to improve error reporting
- Add checks to required flags and tidy up flag handling
- Modify root command to handle flag errors with custom error type
- Remove unused logger and CLI helper files in `pkg/cli`
- Refactor main Go module path to `github.com/intility/minctl`
- Add `pkg/cmderrors` with `InvalidUsageError` for command errors
- Remove the binary files `main` and `minctl` from version control

The commit introduces a `.golangci.yaml` file to the project, enabling a set
of linters to ensure code quality and consistency. It also includes an extensive
refactoring of the cobra commands, improving error handling and removing some
commands in favor of a more streamlined CLI interface. Unnecessary files and
binary outputs are cleaned up from the repository, and the main Go module path
is updated to reflect its location within the `intility` organization. Lastly,
a new error handling strategy is introduced for command-line flags, embedded in
the `pkg/cmderrors` package with a specific `InvalidUsageError` type.
This ensures a streamlined error reporting experience where only the
usage errors prints out the usage help text.